+++ src/libretro/drivers/rumble/__init__.py
@@ -1,2 +1,2 @@
 from .default import *
-from .interface import *
+from .driver import *